feat: add wallet service
