+++ service/impl_miner.go
@@ -1,0 +1,647 @@
+package service
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"reflect"
+
+	"github.com/filecoin-project/go-address"
+	"github.com/filecoin-project/go-bitfield"
+	"github.com/filecoin-project/go-fil-markets/retrievalmarket"
+	"github.com/filecoin-project/go-fil-markets/storagemarket"
+	"github.com/filecoin-project/go-state-types/abi"
+	"github.com/filecoin-project/go-state-types/big"
+	"github.com/filecoin-project/go-state-types/builtin"
+	"github.com/filecoin-project/go-state-types/dline"
+	"github.com/filecoin-project/lotus/chain/actors/builtin/miner"
+	"github.com/filecoin-project/lotus/chain/actors/builtin/power"
+	power2 "github.com/filecoin-project/specs-actors/v2/actors/builtin/power"
+	"github.com/filecoin-project/venus/pkg/constants"
+	"github.com/filecoin-project/venus/venus-shared/actors"
+	lminer "github.com/filecoin-project/venus/venus-shared/actors/builtin/miner"
+	"github.com/filecoin-project/venus/venus-shared/types"
+	msgTypes "github.com/filecoin-project/venus/venus-shared/types/messager"
+	"github.com/google/uuid"
+)
+
+func (s *ServiceImpl) MinerCreate(ctx context.Context, params *MinerCreateReq) (address.Address, error) {
+	// if target msg no exist, send it
+	has := false
+	var err error
+	if params.MsgId != "" {
+		has, err = s.Messager.HasMessageByUid(ctx, params.MsgId)
+		if err != nil {
+			return address.Undef, err
+		}
+	} else {
+		params.MsgId = uuid.New().String()
+	}
+
+	if !has {
+		sealProof, err := miner.SealProofTypeFromSectorSize(params.SectorSize, constants.TestNetworkVersion)
+		if err != nil {
+			return address.Undef, err
+		}
+
+		params.SealProofType = sealProof
+
+		if params.Owner == address.Undef {
+			actor, err := s.Node.StateLookupID(ctx, params.From, types.EmptyTSK)
+			if err != nil {
+				return address.Undef, err
+			}
+			params.Owner = actor
+		}
+
+		if params.Worker == address.Undef {
+			params.Worker = params.Owner
+		}
+
+		p, err := actors.SerializeParams(&params.CreateMinerParams)
+		if err != nil {
+			return address.Undef, err
+		}
+		msg := &types.Message{
+			From:   params.From,
+			To:     power.Address,
+			Method: power.Methods.CreateMiner,
+			Params: p,
+			Value:  big.Zero(),
+		}
+
+		_, err = s.Messager.PushMessageWithId(ctx, params.MsgId, msg, &msgTypes.SendSpec{})
+		if err != nil {
+			return address.Undef, err
+		}
+	}
+
+	ret, err := s.Messager.GetMessageByUid(ctx, params.MsgId)
+	if err != nil {
+		return address.Undef, err
+	}
+
+	switch ret.State {
+	case msgTypes.OnChainMsg, msgTypes.ReplacedMsg:
+		if ret.Receipt.ExitCode != 0 {
+			log.Warnf("message exec failed: %s(%d)", ret.Receipt.ExitCode, ret.Receipt.ExitCode)
+			return address.Undef, fmt.Errorf("message exec failed: %s(%d)", ret.Receipt.ExitCode, ret.Receipt.ExitCode)
+		}
+
+		var cRes power2.CreateMinerReturn
+		err = cRes.UnmarshalCBOR(bytes.NewReader(ret.Receipt.Return))
+		if err != nil {
+			return address.Undef, err
+		}
+
+		return cRes.IDAddress, nil
+
+	case msgTypes.NoWalletMsg:
+		log.Warnf("no wallet available for the sender %s, please check", params.From)
+		return address.Undef, fmt.Errorf("no wallet available for the sender %s, please check", params.From)
+
+	case msgTypes.FailedMsg:
+		log.Warnf("message failed: %s", ret.ErrorMsg)
+		return address.Undef, fmt.Errorf("message failed: %s", ret.ErrorMsg)
+
+	default:
+		log.Infof("msg state: %s", msgTypes.MessageStateToString(ret.State))
+		return address.Undef, fmt.Errorf("temp error: waiting msg (%s) with state(%s) to be on chain", ret.ID, msgTypes.MessageStateToString(ret.State))
+	}
+}
+
+func (s *ServiceImpl) MinerGetStorageAsk(ctx context.Context, mAddr address.Address) (*storagemarket.StorageAsk, error) {
+	sAsk, err := s.Market.MarketGetAsk(ctx, mAddr)
+	if err != nil {
+		return nil, err
+	}
+	return sAsk.Ask, nil
+}
+
+func (s *ServiceImpl) MinerGetRetrievalAsk(ctx context.Context, mAddr address.Address) (*retrievalmarket.Ask, error) {
+	return s.Market.MarketGetRetrievalAsk(ctx, mAddr)
+}
+
+func (s *ServiceImpl) MinerSetStorageAsk(ctx context.Context, p *MinerSetAskReq) error {
+	info, err := s.Node.StateMinerInfo(ctx, p.Miner, types.EmptyTSK)
+	if err != nil {
+		return fmt.Errorf("get miner sector size failed: %s", err)
+	}
+
+	smax := abi.PaddedPieceSize(info.SectorSize)
+
+	if p.MaxPieceSize == 0 {
+		p.MaxPieceSize = smax
+	}
+
+	if p.MaxPieceSize > smax {
+		return fmt.Errorf("max piece size (w/bit-padding) %s cannot exceed miner sector size %s", types.SizeStr(types.NewInt(uint64(p.MaxPieceSize))), types.SizeStr(types.NewInt(uint64(smax))))
+	}
+	return s.Market.MarketSetAsk(ctx, p.Miner, p.Price, p.VerifiedPrice, p.Duration, p.MinPieceSize, p.MaxPieceSize)
+}
+
+func (s *ServiceImpl) MinerSetRetrievalAsk(ctx context.Context, p *MinerSetRetrievalAskReq) error {
+	return s.Market.MarketSetRetrievalAsk(ctx, p.Miner, &p.Ask)
+}
+
+func (s *ServiceImpl) MinerInfo(ctx context.Context, mAddr address.Address) (*MinerInfoResp, error) {
+	mi, err := s.Node.StateMinerInfo(ctx, mAddr, types.EmptyTSK)
+	if err != nil {
+		return nil, fmt.Errorf("get miner(%s) info failed: %s", mAddr, err)
+	}
+	availBalance, err := s.Node.StateMinerAvailableBalance(ctx, mAddr, types.EmptyTSK)
+	if err != nil {
+		return nil, fmt.Errorf("get miner(%s) available balance failed: %s", mAddr, err)
+	}
+
+	power, err := s.Node.StateMinerPower(ctx, mAddr, types.EmptyTSK)
+	if err != nil {
+		return nil, fmt.Errorf("get miner(%s) power failed: %s", mAddr, err)
+	}
+
+	deadline, err := s.Node.StateMinerProvingDeadline(ctx, mAddr, types.EmptyTSK)
+	if err != nil {
+		return nil, fmt.Errorf("get miner(%s) deadline failed: %s", mAddr, err)
+	}
+
+	return &MinerInfoResp{
+		MinerInfo:    mi,
+		MinerPower:   *power,
+		AvailBalance: availBalance,
+		Deadline:     *deadline,
+	}, nil
+}
+
+func (s *ServiceImpl) MinerSetOwner(ctx context.Context, p *MinerSetOwnerReq) error {
+	minerInfo, err := s.Node.StateMinerInfo(ctx, p.Miner, types.EmptyTSK)
+	if err != nil {
+		return fmt.Errorf("get miner(%s) info failed: %s", p.Miner, err)
+	}
+
+	newOwnerId, err := s.Node.StateLookupID(ctx, p.NewOwner, types.EmptyTSK)
+	if err != nil {
+		return fmt.Errorf("get new owner(%s) id failed: %s", p.NewOwner, err)
+	}
+
+	if minerInfo.Owner == newOwnerId {
+		return fmt.Errorf("new owner(%s) is the same as old owner(%s)", p.NewOwner, minerInfo.Owner)
+	}
+
+	param, err := actors.SerializeParams(&newOwnerId)
+	if err != nil {
+		return fmt.Errorf("serialize params failed: %s", err)
+	}
+
+	msg, err := s.PushMessageAndWait(ctx, &types.Message{
+		From:   minerInfo.Owner,
+		To:     p.Miner,
+		Method: builtin.MethodsMiner.ChangeOwnerAddress,
+		Params: param,
+		Value:  big.Zero(),
+	}, nil)
+	if err != nil {
+		return fmt.Errorf("push message(%s) failed: %s", msg.ID, err)
+	}
+
+	return nil
+}
+
+func (s *ServiceImpl) MinerConfirmOwner(ctx context.Context, p *MinerSetOwnerReq) (oldOwner address.Address, err error) {
+	minerInfo, err := s.Node.StateMinerInfo(ctx, p.Miner, types.EmptyTSK)
+	if err != nil {
+		return address.Undef, fmt.Errorf("get miner(%s) info failed: %s", p.Miner, err)
+	}
+	oldOwner = minerInfo.Owner
+
+	newOwnerId, err := s.Node.StateLookupID(ctx, p.NewOwner, types.EmptyTSK)
+	if err != nil {
+		return address.Undef, fmt.Errorf("get new owner(%s) id failed: %s", p.NewOwner, err)
+	}
+
+	if minerInfo.Owner == newOwnerId {
+		return address.Undef, fmt.Errorf("new owner(%s) is the same as old owner(%s)", p.NewOwner, minerInfo.Owner)
+	}
+
+	param, err := actors.SerializeParams(&newOwnerId)
+	if err != nil {
+		return address.Undef, fmt.Errorf("serialize params failed: %s", err)
+	}
+
+	msg, err := s.PushMessageAndWait(ctx, &types.Message{
+		From:   p.NewOwner,
+		To:     p.Miner,
+		Method: builtin.MethodsMiner.ChangeOwnerAddress,
+		Params: param,
+		Value:  big.Zero(),
+	}, nil)
+	if err != nil {
+		return address.Undef, fmt.Errorf("push message(%s) failed: %s", msg.ID, err)
+	}
+
+	return oldOwner, nil
+}
+
+func (s *ServiceImpl) MinerSetWorker(ctx context.Context, req *MinerSetWorkerReq) (WorkerChangeEpoch abi.ChainEpoch, err error) {
+	minerInfo, err := s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return 0, fmt.Errorf("get miner(%s) info failed: %s", req.Miner, err)
+	}
+
+	newWorkerId, err := s.Node.StateLookupID(ctx, req.NewWorker, types.EmptyTSK)
+	if err != nil {
+		return 0, fmt.Errorf("get new worker(%s) id failed: %s", req.NewWorker, err)
+	}
+
+	if minerInfo.Worker == newWorkerId {
+		return 0, fmt.Errorf("new worker(%s) is the same as old worker(%s)", req.NewWorker, minerInfo.Worker)
+	}
+
+	if minerInfo.NewWorker == newWorkerId {
+		return 0, fmt.Errorf("new worker(%s) has been proposed before, which will be effective after epoch(%d)", minerInfo.NewWorker, minerInfo.WorkerChangeEpoch)
+	}
+
+	param, err := actors.SerializeParams(&types.ChangeWorkerAddressParams{
+		NewWorker:       newWorkerId,
+		NewControlAddrs: minerInfo.ControlAddresses,
+	})
+	if err != nil {
+		return 0, fmt.Errorf("serialize params failed: %s", err)
+	}
+
+	msg, err := s.PushMessageAndWait(ctx, &types.Message{
+		From:   minerInfo.Owner,
+		To:     req.Miner,
+		Method: builtin.MethodsMiner.ChangeWorkerAddress,
+		Params: param,
+		Value:  big.Zero(),
+	}, nil)
+	if err != nil {
+		return 0, fmt.Errorf("push message(%s) failed: %s", msg.ID, err)
+	}
+
+	minerInfo, err = s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return 0, fmt.Errorf("get miner(%s) info failed: %s", req.Miner, err)
+	}
+
+	return minerInfo.WorkerChangeEpoch, nil
+}
+
+func (s *ServiceImpl) MinerConfirmWorker(ctx context.Context, req *MinerSetWorkerReq) error {
+	minerInfo, err := s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return fmt.Errorf("get miner(%s) info failed: %s", req.Miner, err)
+	}
+
+	if minerInfo.NewWorker.Empty() {
+		return fmt.Errorf("miner(%s) has no new worker", req.Miner)
+	}
+
+	if minerInfo.NewWorker != req.NewWorker {
+		return fmt.Errorf("new worker(%s) is not the same as proposed worker(%s)", req.NewWorker, minerInfo.NewWorker)
+	}
+
+	head, err := s.Node.ChainHead(ctx)
+	if err != nil {
+		return fmt.Errorf("get chain head failed: %s", err)
+	}
+
+	if head.Height() < minerInfo.WorkerChangeEpoch {
+		return fmt.Errorf("worker change epoch(%d) is not reached", minerInfo.WorkerChangeEpoch)
+	}
+
+	msg, err := s.PushMessageAndWait(ctx, &types.Message{
+		From:   minerInfo.Owner,
+		To:     req.Miner,
+		Method: builtin.MethodsMiner.ConfirmUpdateWorkerKey,
+		Value:  big.Zero(),
+	}, nil)
+	if err != nil {
+		return fmt.Errorf("push message(%s) failed: %s", msg.ID, err)
+	}
+
+	return nil
+}
+
+func (s *ServiceImpl) MinerSetControllers(ctx context.Context, req *MinerSetControllersReq) (oldController []address.Address, err error) {
+	minerInfo, err := s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return nil, fmt.Errorf("get miner(%s) info failed: %s", req.Miner, err)
+	}
+	oldController = minerInfo.ControlAddresses
+
+	newControllers := make([]address.Address, 0, len(req.NewControllers))
+	for _, c := range req.NewControllers {
+		id, err := s.Node.StateLookupID(ctx, c, types.EmptyTSK)
+		if err != nil {
+			return nil, fmt.Errorf("get controller(%s) id failed: %s", c, err)
+		}
+		newControllers = append(newControllers, id)
+	}
+
+	if reflect.DeepEqual(minerInfo.ControlAddresses, newControllers) {
+		return nil, fmt.Errorf("new controllers(%s) is the same as old controllers(%s)", req.NewControllers, minerInfo.ControlAddresses)
+	}
+
+	rawParam := &types.ChangeWorkerAddressParams{
+		NewWorker:       minerInfo.Worker,
+		NewControlAddrs: newControllers,
+	}
+
+	param, err := actors.SerializeParams(rawParam)
+	if err != nil {
+		return nil, fmt.Errorf("serialize params failed: %s", err)
+	}
+
+	msg, err := s.PushMessageAndWait(ctx, &types.Message{
+		From:   minerInfo.Owner,
+		To:     req.Miner,
+		Method: builtin.MethodsMiner.ChangeWorkerAddress,
+		Params: param,
+		Value:  big.Zero(),
+	}, nil)
+	if err != nil {
+		return nil, fmt.Errorf("push message(%s) failed: %s", msg.ID, err)
+	}
+
+	return oldController, nil
+}
+
+func (s *ServiceImpl) MinerSetBeneficiary(ctx context.Context, req *MinerSetBeneficiaryReq) (*types.PendingBeneficiaryChange, error) {
+	minerInfo, err := s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return nil, fmt.Errorf("get miner(%s) info failed: %s", req.Miner, err)
+	}
+
+	newBeneficiary, err := s.Node.StateLookupID(ctx, req.NewBeneficiary, types.EmptyTSK)
+	if err != nil {
+		return nil, fmt.Errorf("get beneficiary(%s) id failed: %s", req.NewBeneficiary, err)
+	}
+	req.NewBeneficiary = newBeneficiary
+
+	if newBeneficiary == minerInfo.Owner {
+		req.NewQuota = big.Zero()
+		req.NewExpiration = 0
+
+		if minerInfo.Beneficiary == newBeneficiary {
+			return nil, fmt.Errorf("beneficiary %s already set to owner address", newBeneficiary)
+		}
+	}
+
+	param, err := actors.SerializeParams(&req.ChangeBeneficiaryParams)
+	if err != nil {
+		return nil, fmt.Errorf("serialize params failed: %s", err)
+	}
+
+	// owner proposal
+	msg, err := s.PushMessageAndWait(ctx, &types.Message{
+		From:   minerInfo.Owner,
+		To:     req.Miner,
+		Method: builtin.MethodsMiner.ChangeBeneficiary,
+		Params: param,
+		Value:  big.Zero(),
+	}, nil)
+	if err != nil {
+		return nil, fmt.Errorf("push message(%s) failed: %s", msg.ID, err)
+	}
+
+	minerInfo, err = s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return nil, fmt.Errorf("get miner(%s) info failed: %s", req.Miner, err)
+	}
+
+	if minerInfo.PendingBeneficiaryTerm == nil {
+		return nil, fmt.Errorf("owner proposal beneficial change failed")
+	}
+
+	return minerInfo.PendingBeneficiaryTerm, nil
+}
+
+func (s *ServiceImpl) MinerConfirmBeneficiary(ctx context.Context, req *MinerConfirmBeneficiaryReq) (confirmor address.Address, err error) {
+	minerInfo, err := s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return address.Undef, fmt.Errorf("get miner(%s) info failed: %s", req.Miner, err)
+	}
+
+	if minerInfo.PendingBeneficiaryTerm == nil {
+		return address.Undef, fmt.Errorf("miner(%s) no pending beneficiary", req.Miner)
+	}
+	if minerInfo.PendingBeneficiaryTerm.NewBeneficiary != req.NewBeneficiary {
+		return address.Undef, fmt.Errorf("new beneficiary(%s) is not the same as proposed beneficiary(%s)", req.NewBeneficiary, minerInfo.PendingBeneficiaryTerm.NewBeneficiary)
+	}
+
+	sender := minerInfo.Beneficiary
+	if !req.ByNominee {
+		if minerInfo.PendingBeneficiaryTerm.ApprovedByBeneficiary {
+			return address.Undef, fmt.Errorf("proposal already approved by beneficiary(%s)", minerInfo.Beneficiary)
+		}
+	} else {
+		if minerInfo.PendingBeneficiaryTerm.ApprovedByNominee {
+			return address.Undef, fmt.Errorf("proposal already approved by nominee(%s)", minerInfo.PendingBeneficiaryTerm.NewBeneficiary)
+		}
+		sender = minerInfo.PendingBeneficiaryTerm.NewBeneficiary
+	}
+
+	param, err := actors.SerializeParams(&types.ChangeBeneficiaryParams{
+		NewBeneficiary: minerInfo.PendingBeneficiaryTerm.NewBeneficiary,
+		NewQuota:       minerInfo.PendingBeneficiaryTerm.NewQuota,
+		NewExpiration:  minerInfo.PendingBeneficiaryTerm.NewExpiration,
+	})
+	if err != nil {
+		return address.Undef, fmt.Errorf("serialize params failed: %s", err)
+	}
+
+	msg, err := s.PushMessageAndWait(ctx, &types.Message{
+		From:   sender,
+		To:     req.Miner,
+		Method: builtin.MethodsMiner.ChangeBeneficiary,
+		Params: param,
+		Value:  big.Zero(),
+	}, nil)
+	if err != nil {
+		return address.Undef, fmt.Errorf("push message(%s) failed: %s", msg.ID, err)
+	}
+
+	return sender, nil
+}
+
+func (s *ServiceImpl) MinerGetDeadlines(ctx context.Context, mAddr address.Address) (*dline.Info, error) {
+	return s.Node.StateMinerProvingDeadline(ctx, mAddr, types.EmptyTSK)
+}
+
+func (s *ServiceImpl) MinerWithdrawToBeneficiary(ctx context.Context, req *MinerWithdrawBalanceReq) (abi.TokenAmount, error) {
+
+	minerInfo, err := s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return big.Zero(), fmt.Errorf("get miner(%s) info failed: %s", req.Miner, err)
+	}
+
+	available, err := s.Node.StateMinerAvailableBalance(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return big.Zero(), fmt.Errorf("get miner(%s) available balance failed: %s", req.Miner, err)
+	}
+
+	if available.LessThan(req.Amount) {
+		return big.Zero(), fmt.Errorf("withdraw amount(%s) is greater than available balance(%s)", req.Amount, available)
+	}
+
+	if req.Amount.LessThanEqual(big.Zero()) {
+		req.Amount = available
+	}
+
+	param, err := actors.SerializeParams(&types.MinerWithdrawBalanceParams{
+		AmountRequested: req.Amount,
+	})
+	if err != nil {
+		return big.Zero(), fmt.Errorf("serialize params failed: %s", err)
+	}
+
+	msg, err := s.PushMessageAndWait(ctx, &types.Message{
+		From:   minerInfo.Beneficiary,
+		To:     req.Miner,
+		Method: builtin.MethodsMiner.WithdrawBalance,
+		Params: param,
+		Value:  big.Zero(),
+	}, nil)
+
+	if err != nil {
+		return big.Zero(), fmt.Errorf("push message(%s) failed: %s", msg.ID, err)
+	}
+
+	return req.Amount, nil
+}
+
+func (s *ServiceImpl) MinerWithdrawFromMarket(ctx context.Context, req *MinerWithdrawBalanceReq) (abi.TokenAmount, error) {
+	minerInfo, err := s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return big.Zero(), fmt.Errorf("get miner(%s) info failed: %s", req.Miner, err)
+	}
+	marketBalance, err := s.Node.StateMarketBalance(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return big.Zero(), fmt.Errorf("get miner(%s) available balance failed: %s", req.Miner, err)
+	}
+
+	reserved, err := s.Market.MarketGetReserved(ctx, req.Miner)
+	if err != nil {
+		return big.Zero(), fmt.Errorf("get miner(%s) reserved balance failed: %s", req.Miner, err)
+	}
+
+	avail := big.Subtract(big.Subtract(marketBalance.Escrow, marketBalance.Locked), reserved)
+
+	if avail.LessThanEqual(big.Zero()) {
+		return big.Zero(), fmt.Errorf("no available balance to withdraw")
+	}
+
+	if avail.LessThan(req.Amount) {
+		return big.Zero(), fmt.Errorf("withdraw amount(%s) is greater than available balance(%s)", req.Amount, avail)
+	}
+
+	if req.Amount.LessThanEqual(big.Zero()) {
+		req.Amount = avail
+	}
+
+	if req.To == address.Undef {
+		req.To = minerInfo.Owner
+	} else {
+		toId, err := s.Node.StateLookupID(ctx, req.To, types.EmptyTSK)
+		if err != nil {
+			return big.Zero(), fmt.Errorf("lookup to address(%s) failed: %s", req.To, err)
+		}
+		req.To = toId
+		if toId != minerInfo.Owner && toId != minerInfo.Worker {
+			return big.Zero(), fmt.Errorf("to address(%s) is not miner owner(%s) or worker(%s)", req.To, minerInfo.Owner, minerInfo.Worker)
+		}
+	}
+
+	mCid, err := s.Market.MarketWithdraw(ctx, req.To, req.Miner, req.Amount)
+	if err != nil {
+		return big.Zero(), fmt.Errorf("withdraw from market failed: %s", err)
+	}
+	id := mCid.String()
+	log.Infof("push message(%s) success", id)
+
+	msg, err := s.MsgWait(ctx, id)
+	if err != nil {
+		return big.Zero(), fmt.Errorf("push message(%s) failed: %s", msg.ID, err)
+	}
+	log.Infof("messager(%s) is chained", id)
+
+	return req.Amount, nil
+}
+
+func (s *ServiceImpl) SectorExtend(ctx context.Context, req SectorExtendReq) error {
+	var err error
+	rawParams := &types.ExtendSectorExpirationParams{}
+
+	sectors := map[lminer.SectorLocation][]abi.SectorNumber{}
+	for _, num := range req.SectorNumbers {
+		p, err := s.Node.StateSectorPartition(ctx, req.Miner, num, types.EmptyTSK)
+		if err != nil {
+			return fmt.Errorf("get sector partition failed: %s", err)
+		}
+
+		if p == nil {
+			return fmt.Errorf("sector %d not found", num)
+		}
+
+		sectors[*p] = append(sectors[*p], num)
+	}
+
+	for p, numbers := range sectors {
+		nums := make([]uint64, len(numbers))
+		for i, n := range numbers {
+			nums[i] = uint64(n)
+		}
+		rawParams.Extensions = append(rawParams.Extensions, types.ExpirationExtension{
+			Deadline:      p.Deadline,
+			Partition:     p.Partition,
+			Sectors:       bitfield.NewFromSet(nums),
+			NewExpiration: req.Expiration,
+		})
+	}
+
+	params, err := actors.SerializeParams(rawParams)
+	if err != nil {
+		return err
+	}
+
+	mi, err := s.Node.StateMinerInfo(ctx, req.Miner, types.EmptyTSK)
+	if err != nil {
+		return fmt.Errorf("get miner info failed: %s", err)
+	}
+
+	_, err = s.PushMessageAndWait(ctx, &types.Message{
+		From:   mi.Worker,
+		To:     req.Miner,
+		Method: builtin.MethodsMiner.ExtendSectorExpiration,
+		Params: params,
+		Value:  big.Zero(),
+	}, &msgTypes.SendSpec{})
+	if err != nil {
+		return fmt.Errorf("push message failed: %s", err)
+	}
+
+	return nil
+}
+
+func (s *ServiceImpl) SectorGet(ctx context.Context, req SectorGetReq) ([]*SectorResp, error) {
+	ret := make([]*SectorResp, 0)
+	for _, num := range req.SectorNumbers {
+		sector, err := s.Node.StateSectorGetInfo(ctx, req.Miner, num, types.EmptyTSK)
+		if err != nil {
+			return nil, fmt.Errorf("get sector(%s) info failed: %s", num, err)
+		}
+
+		p, err := s.Node.StateSectorPartition(ctx, req.Miner, num, types.EmptyTSK)
+		if err != nil {
+			return nil, fmt.Errorf("get sector(%s) partition failed: %s", num, err)
+		}
+
+		ret = append(ret, &SectorResp{
+			SectorOnChainInfo: *sector,
+			SectorLocation:    *p,
+		})
+	}
+
+	return ret, nil
+}
